refactor: extract toast utility
