+++ test/calcDrugs.test.js
@@ -38,7 +38,8 @@ test('calcDrugs handles dosing correctly, validates inputs, and resets outputs',
   };
 
   global.document = documentStub;
-  global.showToast = () => {};
+  const { toast } = await import('../js/toast.js');
+  toast.showToast = () => {};
   global.confirm = () => true;
   global.localStorage = { setItem: () => {}, getItem: () => null };
   global.URL = { createObjectURL: () => '', revokeObjectURL: () => {} };
