+++ test/genSummary.test.js
@@ -39,7 +39,8 @@ test('genSummary generates summary text correctly', async () => {
   };
 
   global.document = documentStub;
-  global.showToast = () => {};
+  const { toast } = await import('../js/toast.js');
+  toast.showToast = () => {};
   global.confirm = () => true;
   global.localStorage = { setItem: () => {}, getItem: () => null };
   global.URL = { createObjectURL: () => '', revokeObjectURL: () => {} };
